Moved images to src folder in react project and created files for GameOption component.
