+++ rock-paper-scissors-master/react-rock-paper-scissors/src/App.js
@@ -1,11 +1,13 @@
 import logo from './logo.svg';
 import './App.css';
-import Header from './components/header';
+import Header from './components/Header/header';
+import GameOption from './components/GameOption/gameOption';
 
 function App() {
   return (
     <div className="App bg-primary.bg-gradient">
       <Header />
+      <GameOption />
     </div>
   );
 }
