+++ rock-paper-scissors-master/react-rock-paper-scissors/src/components/GameOption/gameOption.js
@@ -1,0 +1,7 @@
+import rockIcon from '../../../public/images/icon-rock'
+
+export default function GameOption(props) {
+    return (
+        <img src={rockIcon} />
+    )
+}
