+++ rock-paper-scissors-master/react-rock-paper-scissors/src/images/favicon-32x32.png
